rpi/requirements.txt
pyserial
pyudev